+++ packages/ui/src/views/execution/pipeline-status.tsx
@@ -1,10 +1,17 @@
-import { FC, useEffect, useRef, useState } from 'react'
+import { FC } from 'react'
 
 import { ExecutionState } from '@views/repo/pull-request'
 
 import { ExecutionStatus } from './execution-status'
 
-interface PipelineStatusProps {
+const PipelineStatus = ({
+  status,
+  buildTime,
+  createdTime,
+  commit,
+  branch,
+  delegateType
+}: {
   status: ExecutionState
   buildTime: string
   createdTime?: string
@@ -12,56 +19,7 @@ interface PipelineStatusProps {
   branch?: string
   startedTime?: string
   delegateType?: string
-}
-
-const PipelineStatus: FC<PipelineStatusProps> = ({
-  status,
-  buildTime,
-  createdTime,
-  commit,
-  branch,
-  // startedTime,
-  delegateType
 }) => {
-  const [elapsedTime, setElapsedTime] = useState('00:00')
-  const [createdTimeElapsed, setCreatedTimeElapsed] = useState('00:00')
-  const createdStartRef = useRef<number>(Date.now())
-  const elapsedStartRef = useRef<number>(Date.now())
-
-  const formatTime = (seconds: number) => {
-    const minutes = Math.floor(seconds / 60)
-    const secs = seconds % 60
-    return `${minutes.toString().padStart(2, '0')}:${secs.toString().padStart(2, '0')}`
-  }
-
-  // Created timer (always counts up from 0)
-  useEffect(() => {
-    createdStartRef.current = Date.now()
-
-    const interval = setInterval(() => {
-      const now = Date.now()
-      const totalDiff = Math.floor((now - createdStartRef.current) / 1000)
-      setCreatedTimeElapsed(formatTime(totalDiff))
-    }, 1000)
-
-    return () => clearInterval(interval)
-  }, [])
-
-  // Elapsed timer (stops when status changes to success)
-  useEffect(() => {
-    elapsedStartRef.current = Date.now()
-
-    const interval = setInterval(() => {
-      if (status === 'success') return
-
-      const now = Date.now()
-      const elapsedDiff = Math.floor((now - elapsedStartRef.current) / 1000)
-      setElapsedTime(formatTime(elapsedDiff))
-    }, 1000)
-
-    return () => clearInterval(interval)
-  }, [status])
-
   return (
     <div className="flex justify-between gap-11">
       {commit && (
@@ -82,11 +40,11 @@ const PipelineStatus: FC<PipelineStatusProps> = ({
       </div>
       <div className="flex flex-col gap-1.5">
         <span className="leading-tight text-foreground-4">Build time</span>
-        <span className="text-foreground-1">{elapsedTime}</span>
+        <span className="text-foreground-1">{buildTime}</span>
       </div>
       <div className="flex flex-col gap-1.5">
         <span className="leading-tight text-foreground-4">Created</span>
-        <span className="text-foreground-1">{createdTimeElapsed}</span>
+        <span className="text-foreground-1">{createdTime}</span>
       </div>
       {delegateType && (
         <div className="flex flex-col gap-1.5">
